chore: relocate some classes
