+++ src/main/java/de/stamme/basicquests/model/quests/MineBlockQuest.java
@@ -2,6 +2,7 @@ package de.stamme.basicquests.model.quests;
 
 import de.stamme.basicquests.config.MessagesConfig;
 import de.stamme.basicquests.config.MinecraftLocaleConfig;
+import de.stamme.basicquests.model.rewards.Reward;
 import org.bukkit.Material;
 
 import java.text.MessageFormat;
