+++ src/main/java/de/stamme/basicquests/model/quests/ReachLevelQuest.java
@@ -2,6 +2,7 @@ package de.stamme.basicquests.model.quests;
 
 import de.stamme.basicquests.model.QuestPlayer;
 import de.stamme.basicquests.config.MessagesConfig;
+import de.stamme.basicquests.model.rewards.Reward;
 
 import java.text.MessageFormat;
